+++ paddle/phi/kernels/legacy/gpu/layer_norm_cuda_kernel.cu
@@ -25,12 +25,12 @@ namespace phi {
 
 static void GetRowsCols(const std::vector<int64_t> &shape,
                         int64_t *p_rows,
-                        int *p_cols) {
+                        int64_t *p_cols) {
   int64_t rows = 1;
   for (int i = 0; i + 1 < shape.size(); ++i) {
     rows *= shape[i];
   }
-  int cols = shape[shape.size() - 1];
+  int64_t cols = shape[shape.size() - 1];
   *p_rows = rows;
   *p_cols = cols;
 }
@@ -44,7 +44,7 @@ void RMSLnFwd(const Context &dev_ctx,
               DenseTensor *invvar) {
   const auto &scale_shape = scale.dims();
   int64_t rows;
-  int cols;
+  int64_t cols;
   GetRowsCols(vectorize(x.dims()), &rows, &cols);
   if (scale.dtype() == DataType::BFLOAT16) {
     dev_ctx.template Alloc<bfloat16>(y);
@@ -70,7 +70,7 @@ void RMSLnBwd(const Context &dev_ctx,
               DenseTensor *x_grad,
               DenseTensor *scale_grad) {
   int64_t rows;
-  int cols;
+  int64_t cols;
   GetRowsCols(vectorize(x.dims()), &rows, &cols);
   dev_ctx.template Alloc<T>(x_grad);
   if (scale_grad) {
